board: emulation: Add QEMU sbsa support
Add support for Arm sbsa [1] v0.3+ that is supported by QEMU [2].

Unlike other Arm based platforms the machine only provides a minimal
FDT that contains number of CPUs, ammount of memory and machine-version.
The boot firmware has to provide ACPI tables to the OS.
Due to this design a full DTB is added here as well that allows U-Boot's
driver to properly function. The DTB is appended at the end of the U-Boot
image and will be merged with the QEMU provided DTB.

In addition provide documentation how to use, enable binman to fabricate both
ROMs that are required to boot and add ACPI tables to make it full compatible
to the EDK2 reference implementation.

The board was tested using Fedora 40 Aarch64 Workstation. It's able
to boot from USB and AHCI or network.

Tested and found working:
- serial
- PCI
- xHCI
- Bochs display
- AHCI
- network using e1000e
- CPU init
- Booting Fedora 40

1: Server Base System Architecture (SBSA)
2: https://www.qemu.org/docs/master/system/arm/sbsa.html

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Cc: Peter Robinson <pbrobinson@gmail.com>
Cc: Simon Glass <sjg@chromium.org>
Cc: Tom Rini <trini@konsulko.com>
---
Changelog v3:
- Add GIC and GIC-ITS to devicetree
- Select GICv3 driver
- Drop acpi_fill_madt and use driver model instead
Changelog v4:
- Drop CPU platform code
- Enhance the DT to allow MADT generation from DT
Changelog v5:
- Add full DT and place it at the end of U-Boot
- Merge DT with QEMU's DT
- Drop DT generation code
- Fix flash region length
- Drop enable_caches()
- Support platforms that do not pass FDT in x0
Changelog v6:
- Update header order
- Drop pad-byte from DT
- select BINMAN_FDT
- select E1000_NO_NVM
- drop config.h include
- drop a few CFG_ defines that were used for SPL
Changelog v8:
- Use /bits/ 64 in DT
- Drop flash access helper functions
- Add cfi-flash to DT
- Mark secure-flash as no execute
- Only use defines in DT when it's also used in other files
Changelog v9:
- Use interrupt-cells=3
- Select BLOBLIST
Changelog v10:
- Update defconfig to select BLOBLIST_ALLOC