efi_loader: Install ACPI tables
Install ACPI tables inside the efi_loader similar to SMBIOS tables.
When ACPI is enabled and wasn't installed in other places, install
the ACPI table in EFI. Since EFI is necessary to pass the ACPI
table location when FDT isn't used, there's no need to install it
separately.

When CONFIG_BLOBLIST_TABLES is set the tables will be stored in a
bloblist.

TEST: Booted QEMU SBSA using EFI and ACPI only.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Cc: Simon Glass <sjg@chromium.org>
Cc: Tom Rini <trini@konsulko.com>